#132 Put testing utility to separate module
