+++ ports/cpp/test/utility/Testing.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <gmock/gmock.h>
+#include <gtest/gtest.h>
+
+#include "AntlrPipeline.hpp"
+#include "Maps.hpp"
+
+namespace c3::test {
+
+using testing::ElementsAre;
+using testing::UnorderedElementsAre;
+
+}
