+++ ports/cpp/test/whitebox/WhiteboxTest.cpp
@@ -1,3 +1,6 @@
 #include <gtest/gtest.h>
 
-TEST(HelloWorld, Nothing) {}
+TEST(WhiteboxGrammarTests,
+     CaretAtTransitionToRuleWithNonExhaustiveFollowSetOptionalTokens) {
+  //
+}
